finished core DB functionality
Got all DB methods working.  some returns and parameters were changed.
getLesson returns a single record to be loaded into the program.
getRecord returns an array of all records that match the criteria.  it
will need a simple loop to print all the data to a text file for each
record